+++ scenes/terminal/modals/task_info/task_info_modal.gd
@@ -5,6 +5,7 @@ class_name TaskInfoModal
 @onready var project_display: Label = %ProjectDisplay
 @onready var help_display: RichTextLabel = %HelpDisplay
 @onready var d_hsla_display: dHSLADisplay = %dHSLADisplay
+@onready var task_display: RichTextLabel = %TaskDisplay
 
 @export var task: TaskData
 
@@ -16,6 +17,7 @@ func update():
 	app_display.text = task.app.display_name
 	project_display.text = project_name
 	help_display.text = "[color=purple]%s %s[/color]" % [task.app.command, project_name]
+	task_display.text = task.title
 	
 	d_hsla_display.dHSLA = task.dHSLA
 	d_hsla_display.new_HSLA = Sections.HSLA + task.dHSLA