+++ scenes/terminal/windows/success_display.gd
@@ -1,0 +1,5 @@
+extends Control
+
+
+func _on_close_btn_pressed() -> void:
+	AppManager.close_app()